requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
httpx==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic==1.10.7
openai==0.27.8
